Refactor user authentication to store our user object in the session. Breaks other routes, need to look into it.
