+++ models/userModel.js
@@ -1,4 +1,4 @@
-'use strict';
+    'use strict';
 
 module.exports = function (sequelize, DataTypes) {
     var User = sequelize.define(
@@ -7,7 +7,6 @@ module.exports = function (sequelize, DataTypes) {
             name: DataTypes.STRING,
             authschId: DataTypes.STRING,
             email: DataTypes.STRING,
-            roomNumber: DataTypes.INTEGER,
             admin: DataTypes.BOOLEAN
         },
         {
