+++ routes/groups.js
@@ -9,11 +9,12 @@ var getGroups = require('../middleware/group/getGroups');
 var getGroup = require('../middleware/group/getGroup');
 var joinGroup = require('../middleware/group/joinGroup');
 
-var requireAuthentication = require('../middleware/user/isAuthenticated');
+var requireAuthentication = require('../middleware/user/requireAuthentication');
 
 router.use(requireAuthentication);
 
 router.get('/', getGroups(models), function (req, res) {
+    console.log(req.user);
     res.render('pages/study-groups/study-groups', {
         userData: req.user,
         groups: req.group,
