+++ views/layout/main.ejs
@@ -39,11 +39,11 @@
                     <!-- SIDEBAR USER TITLE -->
                     <div class="profile-usertitle">
                         <div class="profile-usertitle-name">
-                            <%= userData.displayName %>
+                            <%= userData.name %>
                         </div>
-                       <!-- <div class="profile-usertitle-job">
-                            <%= userData.roomNumber %>
-                        </div>-->
+                       <div class="profile-usertitle-job">
+                            <%= userData.email %>
+                        </div>
                     </div>
                     <% } %>
                     <!-- END SIDEBAR USER TITLE -->
